Продолжение разработки интерфейса для конструкторов, подключение и добавление данных в листбокс

+++ DatabaseManagement/MainDb/Employee.cs
@@ -19,7 +19,7 @@ public partial class Employee
 
     public long? Phonenum { get; set; }
 
-    public int? Passport { get; set; }
+    public long? Passport { get; set; }
 
     public virtual Designer? Designer { get; set; }
